requirements.txt
Django==4.1.4
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-filter==22.1
sqlparse==0.4.2
djangorestframework-simplejwt==5.0
django-templated-mail==1.1.1
requests==2.28.1
python-dotenv

